backend/requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
Authlib==1.6.1
alembic
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.0
fastapi-limiter==0.1.6
google-auth==2.40.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
numpy==2.2.6
pytest-asyncio
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyotp==2.9.0
pytest==8.4.1
python-dotenv==1.1.1
python-jose==3.5.0
redis>=6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
itsdangerous
prometheus_fastapi_instrumentator
pydantic-settings
requests
geoip2
python-multipart
asgi-lifespan
